LayoutTests/http/tests/media/media-stream/get-user-media-in-embed-element-expected.txt

CONSOLE MESSAGE: Permission policy 'Camera' check failed for element with origin 'http://localhost:8000' and allow attribute ''.
CONSOLE MESSAGE: Not allowed to call getUserMedia.


PASS Same origin embed should get access to camera
PASS Cross origin embed should not get access to camera
